feat: allow customising context via callback on transport servers (#32)
* feat: allow customising context via callback on transport servers

This commit adds an callback function to the transport servers that allows
developers to inject context values into the server context.

This can be used to inject context values extracted from environment variables
(in stdio mode) or from headers (in sse mode), and access them in tools
using the provided context.

* Add tests for custom context functions

* Use t.Setenv instead of manual os.Setenv

* Use existing Option type from main branch for configuring SSE server with context func

* Add StdioOption to configure context func on Stdio server

* Add example of customising context