fix(tools): add `omitempty` to properties (#116)
When there is no input to a tool, the `Properies` map is `nil` if the
Tool is not created using `NewTool`. This causes the `property` field to
be set to `null` and clients such as `claude-desktop` to skip the tools
silently.

Annotating the field as `omitempty` is inline with the behavior of the
[typescript-sdk](https://github.com/modelcontextprotocol/typescript-sdk/blob/fbdeb06a4185ba2f7581603a768cd7171c64d6fc/src/types.ts#L749).