feat(server): Add hooks.AddOnUnregisterSession functionality (#175)
Add OnUnregisterSession hook functionality to complement the existing 
OnRegisterSession hooks, allowing code to run when a client session 
is being removed from the server. In some cases, the server may want to
do additional work when a session has been closed. For example, in the
SSE server case where you may end up managing various logs for the
duration of the session -- you would want to indicate that the session
was finished.