Add `mcptest` package for in-process MCP testing (#149)
* feat(transport): Add the `NewIO` function.

This function allows creating a `*transport.Stdio` using provided
`io.Reader` and `io.Writer`. This allows creating an MCP client to a
server running in the same process, which significanly simplifies
testing.

* feat(mcptest): Add package help testing.

The new `mcptest` package provides functions for setting up an in-process MCP
server and an MCP client connected to it. This allows testing MCP
implementations end-to-end without spawning any processes.

The `mcptest` package itself has a unit test that demonstrates how to use the
package.

* feat(mcptest): Return an initialized client.

This allows to omit the initialization code in tests, making them less verbose.

* fix(mcptest): Close pipes on shutdown.

* refactor(transport): Rename `startProc` to `spawnCommand`.

The new name is more descriptive.